+++ src/pages/search/selector/SearchAreas.tsx
@@ -6,7 +6,7 @@ import {
   type LegacyContextValues,
 } from "pages/search/SearchContext";
 import { AreaIdBasic, AreaType } from "pages/search/types/dashboard";
-import SearchAPI from "pages/search/utils/searchAPI";
+import SearchAPI from "pages/search/api/searchAPI";
 
 interface SearchAreasProps {
   areasList: Array<AreaType>;
